Windows support

+++ cppbuild.bash
@@ -3,11 +3,17 @@ pushd .
 mkdir cppbuild
 cd cppbuild
 
-wget -nc https://dn-cap.com/CapturyLiveSDK.tar.gz
-tar -xvf CapturyLiveSDK.tar.gz
+if [ ! -f "CapturyLiveSDK.tar.gz" ]; then
+  curl -O https://dn-cap.com/CapturyLiveSDK.tar.gz
+fi
+
+#tar -xvf CapturyLiveSDK.tar.gz
 
 cp ../CMakeLists.txt.RemoteCaptury sdk/network/CMakeLists.txt
 cp ../RemoteCaptury.cpp.patch sdk/network/RemoteCaptury.cpp.patch
+
+# Convert all line endings to unix
+find ./sdk -type f -print0 | xargs -0 dos2unix
 
 cd sdk/network
 
@@ -15,9 +21,11 @@ patch RemoteCaptury.cpp RemoteCaptury.cpp.patch
 
 mkdir build
 cd build
-cmake ..
-make
-make install
+
+cmake -DCMAKE_BUILD_TYPE=Release ..
+cmake --build . --config Release
+cmake --build . --target install
+
 popd
 #### Java generation ####
 cd cppbuild
@@ -37,5 +45,17 @@ cp us/ihmc/remotecaptury/global/*.java ../src/main/java/us/ihmc/remotecaptury/gl
 java -jar javacpp.jar us/ihmc/remotecaptury/*.java us/ihmc/remotecaptury/global/*.java -d javainstall
 
 #### Copy shared libs to resources ####
-cp javainstall/libjniremotecaptury.so ../src/main/resources/native/linux-x86_64
-cp lib/libRemoteCaptury.so ../src/main/resources/native/linux-x86_64
+# Linux
+if [ -f "javainstall/libjniremotecaptury.so" ]; then
+  cp javainstall/libjniremotecaptury.so ../src/main/resources/native/linux-x86_64
+fi
+if [ -f "lib/libRemoteCaptury.so" ]; then
+  cp lib/libRemoteCaptury.so ../src/main/resources/native/linux-x86_64
+fi
+# Windows
+if [ -f "javainstall/jniremotecaptury.dll" ]; then
+  cp javainstall/jniremotecaptury.dll ../src/main/resources/native/windows-x86_64
+fi
+if [ -f "bin/RemoteCaptury.dll" ]; then
+  cp bin/RemoteCaptury.dll ../src/main/resources/native/windows-x86_64
+fi
